Few bugs removed
The GameState->Winning() function can't tell who is winning person and don't check winning move for the opposite. Structure of game isn't good and we can't tell only from move that is it opposite player or not.

TODO: After apply move the turning player is changing, we don't want that in our algo.